feat: usePulseAnimation 훅 생성
